+++ ppx/mapper.ml
@@ -93,18 +93,7 @@ let new_landmark landmark_name loc =
 
 let qualified ctx name = String.concat "." (List.rev (name :: ctx))
 
-let raise_ident =
-  let stdlib_raise = "Stdlib.raise" in
-  try
-    let major, minor =
-      Scanf.sscanf Sys.ocaml_version "%d.%d" (fun major minor -> major, minor)
-    in
-    if major = 4 && minor < 7 then
-      "Pervasives.raise"
-    else
-      stdlib_raise
-  with Scanf.Scan_failure _ -> stdlib_raise
-
+let raise_ident = "Landmark.raise"
 
 let wrap_landmark ctx landmark_name loc expr =
   let landmark_name = qualified ctx landmark_name in