+++ CHANGES.md
@@ -4,6 +4,7 @@ version 1.3, 6 dec 2018
 * migrate to opam 2.0
 * adds two new primitives: push/pop_profiling_state
 * prepare migration for Pervasives deprecation
+* redefine 'Stdlib.raise' in the 'Landmark' to allow using landmark with -no-stdlib (as 'raise' is used to wrap exception in the generated code).
 
 version 1.2, 20 may 2018
 ------------------------